SalarioBeneficioController

+++ src/main/java/com/cicvieira/apisspringboot/repositoy/CargoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author cicero.silva on 05/12/2022
  * @project apis-spring-boot
  */
-public interface CargoRepository extends JpaRepository<Cargo, Integer> {
+public interface CargoRepository extends JpaRepository<Cargo, Long> {
 
 
 }
