+++ src/main/java/com/cicvieira/apisspringboot/repositoy/VendaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author cicero.silva on 05/12/2022
  * @project apis-spring-boot
  */
-public interface VendaRepository extends JpaRepository<Venda, Integer> {
+public interface VendaRepository extends JpaRepository<Venda, Long> {
 
 
 }
